refactor(entities): nest the bilingual endpoint as a resource
entities.create-bilingual is not a canonical capability verb; the handler
moves to entities/bilingual/create.ts so the id reads entities.bilingual.create.
Regenerates the capability catalog.
